Update HollisT_Week7.Rmd
Assignment:

A researcher hypothesizes that TV stimulation for kids will result in an increase in the number of pieces of candy corn eaten. The kids are subjected to 5 hours of television. The number of candy corn candies consumed during a 10-minute period is counted both with and without TV. Compute the appropriate t-test for the data provided below.



Experiment #

With TV

Without TV

1

12

8

2

7

7

3

3

4

4

11

14

5

8

6

6

5

7

7

14

12

8

7

5

9

9

5

10

10

8



What is your computed answer for the following, based on the date in the table above?




Mean =

Mean =

S =

S =

S2 =

S2 =



What is your computed t-test answer?
What is the null hypothesis in this study?
What is the alternate hypothesis?
What probability level did you choose and why?
What are your degrees of freedom?
Is there a significant difference between the two testing conditions? Interpret your answer.


```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE,
  fig.align = 'center',
  fig.width = 6,
  fig.height = 4,
  out.width = '80%'
  )

library(knitr)
library(kableExtra)

options(
  digits = 2,
  scipen = 999
)
```


## **TV Stimulation and Candy Corn Consumption**

### **Hypothesis**

A researcher hypothesizes that TV stimulation increases candy corn consumption
in children. Each child observed under two conditions, with and without 5 hours
of TV, and the number of candies eaten in 10 minutes is recorded. We will first
calculate the mean, standard deviation, and variance of both vectors.