Update HollisT_Week5.py
The purpose of this week’s assignment is to demonstrate that you can utilize loops in a Python program. You must demonstrate that you can create a while loop which will allow a user to continue the execution of the program until a condition is met which will exit the execution. You must also demonstrate that you can use a for loop to perform an average calculation.

You will create a new program which uses a total of 2 functions and a main method. One of these functions will perform various mathematical calculations (+,-,/,*). The other function will perform an average calculation based upon a series of numbers entered by the user. Your main method will serve as the entry point to the program and will contain a while loop which will allow the user to continue running the program (performing a mathematical calculation or an average calculation) until they wish to exit.



Your program must have the following elements:
Your program must have a header.
Try blocks where appropriate to prevent unhandled exceptions.
This program will perform various calculations (addition, subtraction, multiplication, division, and average calculation).
This program will contain a variety of loops and functions.
The program will add, subtract, multiply, divide two numbers and provide the average of multiple numbers input by the user.
Define a function named perform_calculation which takes one parameter. The parameter will be the operation being performed (+, -, *, /).
This function prompts the user for two numbers and then performs the specified operation based on the parameter passed into the function.
This function will return the calculated value to main where the calculated value will be printed in a legible and contextual format for the end user.
Define a function named calculate_average which takes no parameters.
This function will ask the user how many numbers they wish to input.
This function will use the number entered by the user to control a for loop. Within the for loop you must prompt the user for a numeric value which will be used to calculate the total and average.
This function will return the calculated average to main where it will be printed in a legible and contextual format for the end user.
This program will have a main method which contains a while loop. The while loop will be used to allow the user to run the program until they enter a value which ends the loop.
The main method should prompt the user for the operation they wish to perform (average or calculation).
The main method should evaluate the entered data using if statements.
The main method should call the necessary function to perform the calculation or find the average.
Must use good coding practices.