Update HollisT_Week6.py
The purpose of this week’s assignment is to demonstrate the use of strings and loops in a Python program. This week we will create a program which works with lists. Your goal is to create a program which contains a list of temperatures. Your program will populate the list by prompting the user to input temperatures. Your program will determine the number of temperatures entered by the user and stored in the list, determine the largest temperature in the list, and the smallest temperature in the list.

Your program must have the following elements:
A header.
Try blocks where appropriate to prevent unhandled exceptions.
A Main method.
An empty list called temperatures.
Allow the user to input a series of temperatures along with a sentinel value which will stop the user input.
Evaluate the temperatures in the temperatures list to determine the largest and smallest temperature.
Print the largest temperature in a legible format.
Print the smallest temperature in a legible format.
Print a message that tells the user how many temperatures are in the list.
Must use good coding practices.