Update HollisT_Week7.py
The purpose of this week’s assignment is to demonstrate that you can work with creating a dictionary by processing a file to parse the content and store it in a dictionary. We will create a program which performs three operations to process a file. It will open and process a txt file (gettysburg.txt) to create a dictionary of key value pairs. The key will be the unique word and the value will be the number of occurrences of that word in the file. To determine the unique words your program must remove punctuation from the words, convert the words to a common case, and remove any whitespace.



Your program should open the file and process each line.
Add each word to the dictionary with a frequency of 1 or update the word’s count by 1 if the word already exists in the dictionary.
Nicely print the output in a tabular format with a header (word/count) using string formatting (format() or f-strings) in sorted order from high to low frequency.
We want to achieve each major goal with a function (one function, one action). We can find four functions that need to be created.



add_word: Add each word to the dictionary. Parameters are the word and a dictionary.

This function will not have a return value.

process_line: There is some work to be done to process the line: strip off various characters, split out the words, remove punctuation. Parameters for this function are a line and the dictionary. This function will call add_word with each processed word to increment the word count. This function will not have a return value.

pretty_print: Because unformatted printing can be messy and often particular to each situation (meaning that we might need to modify it later), we separated out the printing function. The parameter for this function is the dictionary. The function will not have a return value. This function must use string formatting to print the output in a tabular format.

main: We will use the main method as the entry point to the program. Main will open the file and call process_line on each line. To allow your program to open the Gettysburg.txt file place the Gettysburg.txt file in the same directory as your program. The main method should print the total length of the dictionary (see image). When finished, it will call pretty_print to print dictionary in a tabular format.



In main you will need to open the file. We will cover more regarding the opening of files next week but I wanted to provide you with the block of code you will utilize to open the file, see below.


Python code attempts to open and read "gettysburg.txt" line by line, passing each line to the process_line function with word_count_dict. If the file is not found, it catches a FileNotFoundError and prints the error message.


Example Output from program execution (note the word count i.e. length of the dictionary):



A word frequency table showing the top words from a text. The dictionary has 143 unique words. The most frequent word is "that" with 13 occurrences, followed by "the" with 11, and "we" with 10. The table displays each word alongside its count in descendin


Your program must have the following elements:
A header.
Try blocks where appropriate to prevent unhandled exceptions.
A Main method.
The following functions add_word, process_line, pretty_print which includes the logic discussed above.
A dictionary which stores the unique word and its number of occurrences within the file.
A print statement which prints the length of the dictionary.
Do not hard code a path to the gettysburg.txt file as this may break on other user’s workstations.
Must use good coding practices.