Update HollisT_Week8.py
The purpose of this week’s assignment is to demonstrate that you can process a file as well as write data to a new file. Last week we got a taste of working with files and we’ll expand on that this week.



For this week we will modify our Gettysburg processing program from Week 7 to generate a text file from the output rather than printing it to the screen. Your program should have a new function called process_file which prints to the new file. This method should almost be the same as the pretty_print function from last week. Keep in mind that we have print statements in main as well. Your program must modify the print statements from main so that they write to the file.



In addition to the Week 8 functionality your program must have the following elements:
A new function named process_file. This function will perform the same operations as pretty_print from Week 7 however it will print to a file instead of to the screen.
A modified main method to print the length of the dictionary to the file as opposed to the screen.
Since you are now writing to a file in two locations (main and process_file you will need to open the file twice. Once in main and once in process_file.
Prompt the user for the filename they wish to use to generate the new file.
Use the filename specified by the user to write create and write to the new file.
This will require you to pass the file as an additional parameter to your new process_file function.
Your program must have try blocks where appropriate including to catch file handling exceptions.
Must use good coding practices.